+++ pkg/databases/migrations/000001_init_db.up.sql
@@ -25,7 +25,7 @@ CREATE TABLE "users" (
   "password" VARCHAR NOT NULL,
   "avatar_url" VARCHAR NOT NULL DEFAULT '',
   "is_admin" BOOLEAN NOT NULL DEFAULT FALSE,
-  "coin" INT NOT NULL DEFAULT 0,
+  "coin" FLOAT NOT NULL DEFAULT 0,
   "created_at" TIMESTAMP DEFAULT now(),
   "updated_at" TIMESTAMP DEFAULT now()
 );
